+++ src/test/java/kr/hanjari/backend/domain/club/domain/entity/ClubCategoryTest.java
@@ -10,8 +10,10 @@ import kr.hanjari.backend.domain.club.domain.enums.College;
 import kr.hanjari.backend.domain.club.domain.enums.Department;
 import kr.hanjari.backend.domain.club.domain.enums.RecruitmentStatus;
 import kr.hanjari.backend.domain.club.domain.enums.UnionClubCategory;
+import kr.hanjari.backend.domain.club.presentation.dto.request.CategoryRequest;
 import kr.hanjari.backend.domain.club.presentation.dto.request.ClubBasicInformationRequest;
-import kr.hanjari.backend.domain.club.presentation.dto.request.ClubBasicInformationRequest.Category;
+import kr.hanjari.backend.domain.club.presentation.dto.request.ClubBasicInformationUpdateRequest;
+import kr.hanjari.backend.domain.club.presentation.dto.util.CategoryUtils;
 import kr.hanjari.backend.domain.common.command.CategoryCommand;
 import kr.hanjari.backend.global.payload.exception.GeneralException;
 import org.junit.jupiter.api.DisplayName;
@@ -40,20 +42,18 @@ class ClubCategoryTest {
     @DisplayName("CENTRAL: centralCategory 설정 및 나머지 필드 초기화")
     void centralCategory_set_and_clear_others() {
         Club club = newClub();
-
-        ClubBasicInformationRequest dto = new ClubBasicInformationRequest(
-                "보안동아리",
+        CategoryRequest categoryRequest = new CategoryRequest(CentralClubCategory.ACADEMIC,
+            null,
+            null,
+            null);
+        ClubBasicInformationUpdateRequest dto = new ClubBasicInformationUpdateRequest(
                 "lead@univ.ac.kr",
                 ClubType.CENTRAL,
-                new Category(CentralClubCategory.ACADEMIC,
-                        null,
-                        null,
-                        null),
-                "보안 좋아함",
-                "CTF 스터디");
+                categoryRequest,
+                "보안 좋아함");
 
-        dto.validate();
-        CategoryCommand cmd = dto.toCategoryCommand();
+        CategoryUtils.validate(ClubType.CENTRAL, categoryRequest);
+        CategoryCommand cmd = CategoryUtils.toCategoryCommand(ClubType.CENTRAL, categoryRequest);
 
         club.updateClubCommonInfo(dto, cmd);
 
@@ -70,18 +70,17 @@ class ClubCategoryTest {
     @DisplayName("UNION: unionCategory 설정 및 나머지 필드 초기화")
     void unionCategory_set_and_clear_others() {
         Club club = newClub();
-
-        ClubBasicInformationRequest dto = new ClubBasicInformationRequest(
+        CategoryRequest categoryRequest = new CategoryRequest(null, UnionClubCategory.MARKETING_AD,
+            null, null);
+        ClubBasicInformationUpdateRequest dto = new ClubBasicInformationUpdateRequest(
                 "연합마케팅",
-                "lead@univ.ac.kr",
                 ClubType.UNION,
-                new Category(null, UnionClubCategory.MARKETING_AD, null, null),
-                "실전 마케팅",
-                "프로모션 실습"
+                categoryRequest,
+                "실전 마케팅"
         );
 
-        dto.validate();
-        CategoryCommand cmd = dto.toCategoryCommand();
+        CategoryUtils.validate(ClubType.UNION, categoryRequest);
+        CategoryCommand cmd = CategoryUtils.toCategoryCommand(ClubType.UNION, categoryRequest);
 
         club.updateClubCommonInfo(dto, cmd);
 
@@ -96,19 +95,17 @@ class ClubCategoryTest {
     @DisplayName("COLLEGE: college 설정(학과 없음) 및 나머지 필드 초기화")
     void college_only_set_and_clear_others() {
         Club club = newClub();
-
-        ClubBasicInformationRequest dto = new ClubBasicInformationRequest(
+        CategoryRequest categoryRequest = new CategoryRequest(null, null, College.ENGINEERING,
+            null);
+        ClubBasicInformationUpdateRequest dto = new ClubBasicInformationUpdateRequest(
                 "공대 세미나",
-                "lead@univ.ac.kr",
                 ClubType.COLLEGE,
-                new Category(null, null, College.ENGINEERING, null),
-                "공대 전체 세미나",
-                "격주 세미나"
+                categoryRequest,
+                "공대 전체 세미나"
         );
 
-        dto.validate();
-        CategoryCommand cmd = dto.toCategoryCommand();
-
+        CategoryUtils.validate(ClubType.COLLEGE, categoryRequest);
+        CategoryCommand cmd = CategoryUtils.toCategoryCommand(ClubType.COLLEGE, categoryRequest);
         club.updateClubCommonInfo(dto, cmd);
 
         assertEquals(ClubType.COLLEGE, club.getCategoryInfo().getClubType());
@@ -122,18 +119,20 @@ class ClubCategoryTest {
     @DisplayName("DEPARTMENT: college + department 설정 및 관계 검증 통과")
     void department_set_with_valid_relation() {
         Club club = newClub();
+        CategoryRequest categoryRequest = new CategoryRequest(null, null, College.ENGINEERING,
+            Department.MECHANICAL);
 
-        ClubBasicInformationRequest dto = new ClubBasicInformationRequest(
+        ClubBasicInformationUpdateRequest dto = new ClubBasicInformationUpdateRequest(
                 "기계과 CAD 연구회",
-                "lead@univ.ac.kr",
                 ClubType.DEPARTMENT,
-                new Category(null, null, College.ENGINEERING, Department.MECHANICAL),
-                "CAD 마스터",
-                "설계 툴 스터디"
+                categoryRequest,
+                "CAD 마스터"
         );
 
-        dto.validate();
-        CategoryCommand cmd = dto.toCategoryCommand();
+
+        CategoryUtils.validate(ClubType.DEPARTMENT, categoryRequest);
+
+        CategoryCommand cmd = CategoryUtils.toCategoryCommand(ClubType.DEPARTMENT, categoryRequest);
 
         club.updateClubCommonInfo(dto, cmd);
 
@@ -148,19 +147,19 @@ class ClubCategoryTest {
     @DisplayName("DEPARTMENT: 학과-단과대 불일치 시 예외")
     void department_set_with_invalid_relation_should_throw() {
         Club club = newClub();
-
+        CategoryRequest categoryRequest = new CategoryRequest(null, null, College.SOFTWARE,
+            Department.MECHANICAL);
         // 가정: Department.MECHANICAL 은 College.ENGINEERING 소속
-        ClubBasicInformationRequest dto = new ClubBasicInformationRequest(
-                "잘못된 소속 테스트",
+        ClubBasicInformationUpdateRequest dto = new ClubBasicInformationUpdateRequest(
                 "lead@univ.ac.kr",
                 ClubType.DEPARTMENT,
-                new Category(null, null, College.SOFTWARE, Department.MECHANICAL),
-                "invalid",
+                categoryRequest,
                 "invalid"
         );
 
-        dto.validate();
-        CategoryCommand cmd = dto.toCategoryCommand();
+        CategoryUtils.validate(ClubType.DEPARTMENT, categoryRequest);
+
+        CategoryCommand cmd = CategoryUtils.toCategoryCommand(ClubType.DEPARTMENT, categoryRequest);
 
         assertThrows(GeneralException.class, () -> club.updateClubCommonInfo(dto, cmd));
     }
@@ -173,28 +172,30 @@ class ClubCategoryTest {
                 "연합동아리",
                 "lead@univ.ac.kr",
                 ClubType.UNION,
-                new Category(CentralClubCategory.ACADEMIC, null, null, null),
+                new CategoryRequest(CentralClubCategory.ACADEMIC, null, null, null),
                 "bad",
                 "bad"
         );
 
-        assertThrows(IllegalArgumentException.class, bad::validate);
+//        assertThrows(IllegalArgumentException.class, CategoryUtils.validate(ClubType.UNION, bad.category()));
     }
 
     @Test
     @DisplayName("DTO → CategoryCommand 변환 확인")
     void dto_to_command() {
+        CategoryRequest categoryRequest = new CategoryRequest(null, null, College.ENGINEERING,
+            Department.MECHANICAL);
         ClubBasicInformationRequest dto = new ClubBasicInformationRequest(
                 "기계과 CAD 연구회",
                 "lead@univ.ac.kr",
                 ClubType.DEPARTMENT,
-                new Category(null, null, College.ENGINEERING, Department.MECHANICAL),
+                categoryRequest,
                 "CAD",
                 "설계"
         );
-        dto.validate();
+        CategoryUtils.validate(ClubType.DEPARTMENT, categoryRequest);
 
-        CategoryCommand cmd = dto.toCategoryCommand();
+        CategoryCommand cmd = CategoryUtils.toCategoryCommand(ClubType.DEPARTMENT, categoryRequest);
         assertEquals(ClubType.DEPARTMENT, cmd.type());
         assertNull(cmd.central());
         assertNull(cmd.union());
